refactor: deleted scheduler_base.rs
